Removed apiCall.
This commit removes apiCall as a distinct field. Instead it
puts the apiCall into the trace field. It is separated from the stack
trace of the user's code by an @ symbol. This makes the trace
look very similar to what Fluid currently produces. TraceTest is updated
and shows the new behavior.